Island: Implement `IAgentPluginService.install_plugin_archive`
